+++ xlsxreport/scripts/compile_excel.py
@@ -1,3 +1,4 @@
+"""Command to compile a formatted Excel from a csv file and a formatting template."""
 import os
 import warnings
 
@@ -24,42 +25,48 @@ HELP = {
 
 
 @click.command()
-@click.argument("infile", type=click.Path(exists=True))
+@click.argument("infile", type=click.Path(exists=True, readable=True))
 @click.argument("template")
-@click.option("--outfile", default="", help=HELP["outfile"])
-@click.option("--outpath", default="", help=HELP["outpath"])
+@click.option("--outfile", help=HELP["outfile"])
+@click.option(
+    "--outpath", help=HELP["outpath"], type=click.Path(exists=True, writable=True)
+)
 @click.option("--sep", default="\t", help=HELP["sep"])
-def report(infile: str, template: str, outfile: str, outpath: str, sep: str) -> None:
-    """Create a formatted Excel report from csv INFILE and a formatting TEMPLATE file.
+def compile_excel(
+    infile: str, template: str, outfile: str, outpath: str, sep: str
+) -> None:
+    """Create a formatted Excel report from a csv INFILE and a formatting TEMPLATE file.
 
     The TEMPLATE argument is first used to look for a file with the specified filepath. If
     no file is found, the XlsxReport appdata directory is searched for a file with the
     corresponding name.
+
+    Args:
+        infile: Path to the input csv file.
+        template: Path to the formatting template file.
+        outfile: Name of the report file, by default the INFILE name is used with the
+            file extension replaced by '.report.xlsx'.
+        outpath: Output path of the report file. If specified overrides the `outfile`
+            option.
+        sep: Delimiter to use for the input file, default is \\t.
+
+    Raises:
+        click.ClickException: If the template file is not found.
+        click.ClickException: If the output path directory does not exist.
     """
     template_path = xlsxreport.appdir.get_template_path(template)
     if template_path is None:
         raise click.ClickException(
-            f"Invalid value for `TEMPLATE`: '{template}' not found."
+            f"Invalid value for `template`: '{template}' file not found."
         )
 
-    if outpath:
-        if not os.path.isdir(os.path.dirname(outpath)):
-            outdir = os.path.dirname(outpath)
-            raise click.ClickException(
-                f"Invalid value for `outpath`: '{outdir}' directory not found."
-            )
-        report_path = outpath
-    elif outfile:
-        report_path = os.path.join(os.path.dirname(infile), outfile)
-    else:
-        infilename = os.path.basename(infile)
-        outfilename = ".".join(infilename.split(".")[:-1]) + ".report.xlsx"
-        report_path = os.path.join(os.path.dirname(infile), outfilename)
+    report_path = _get_report_output_path(infile, outfile, outpath)
 
     click.echo(f"Generating formatted Excel report:")
     click.echo(f"----------------------------------")
     click.echo(f"Input file:    {infile}")
     click.echo(f"Template file: {template_path}")
+    click.echo(f"Report file:   {report_path}")
 
     with warnings.catch_warnings():
         warnings.simplefilter(action="ignore", category=pd.errors.DtypeWarning)
@@ -74,34 +81,16 @@ def report(infile: str, template: str, outfile: str, outpath: str, sep: str) -> 
         section_writer.write_sections(
             worksheet, table_sections, settings=report_template.settings
         )
-    click.echo(f"Report file:   {report_path}")
 
 
-@click.command()
-def setup() -> None:
-    """Setup app directory and copy default template files."""
-    data_dir = xlsxreport.appdir.locate_appdir()
-    if os.path.isdir(data_dir):
-        click.echo(f"App data directory for XlsxReport found at:")
-        click.echo(f"  {data_dir}")
+def _get_report_output_path(infile: str, outfile: str, outpath: str) -> str:
+    """Return the path for the Excel report file."""
+    if outpath:
+        report_path = outpath
+    elif outfile:
+        report_path = os.path.join(os.path.dirname(infile), outfile)
     else:
-        click.echo(f"Creating XlsxReport folder in the local user data directory at:")
-        click.echo(f"  {data_dir}")
-    click.echo(
-        "Copying default XlsxReport template files to the app data directory ..."
-    )
-    xlsxreport.appdir.setup_appdir(overwrite_templates=True)
-    click.echo(f"  Template files were successfully copied.")
-
-
-@click.group()
-def cli():
-    pass
-
-
-cli.add_command(report)
-cli.add_command(setup)
-
-
-if __name__ == "__main__":
-    cli()
+        infilename = os.path.basename(infile)
+        outfilename = ".".join(infilename.split(".")[:-1]) + ".report.xlsx"
+        report_path = os.path.join(os.path.dirname(infile), outfilename)
+    return report_path